Replace Texture- and DrawablesDrawer w/ functions
Add functions draw_texture and draw_drawables to Draw as replacement
Make draw module private and re-export types
Move Game, LogicArgs and RenderArgs into new module game, and re-export them
Minor clean-up on Game
Update example basic
